LayoutTests/http/tests/navigation/back-in-iframe-does-not-stop-loading.html

<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}

window.onmessage = function (event) {
    console.log(`Received event: ${event.data}`);
    if (window.testRunner) {
        testRunner.notifyDone();
    }
}
</script>

<p>This checks that history.back() in an iframe does not stop outstanding network requests.</p>
<iframe src="resources/back-in-iframe-does-not-stop-loading-iframe.html"></iframe>
