LayoutTests/http/tests/navigation/back-in-iframe-does-not-stop-loading-expected.txt

CONSOLE MESSAGE: Received event: fetch completed
This checks that history.back() in an iframe does not stop outstanding network requests.
